requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
click==8.0.1
cmake==3.18.4.post1
dataclasses==0.8
Django==3.2.3
djangorestframework==3.12.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.37.1
h5py==3.1.0
importlib-metadata==4.0.1
imutils==0.5.4
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.0
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
